Removed functions that could be replaced with deep merge. Removed old tests since most of them are irrelevent.
