+++ pyproject.toml
@@ -41,7 +41,8 @@ classifiers = [
     'Topic :: Internet',
 ]
 dependencies = [
-	"pydantic",
+	"pydantic>=2",
+  "pydantic-settings>=2",
 	"pyyaml",
 ]
 requires-python = ">=3.7" # because https://github.com/pydantic/pydantic/blob/main/pyproject.toml
@@ -53,8 +54,8 @@ ci = [
   "twine",
 ]
 dev = [
-	"black", 
-	"flake8", 
+	"black",
+	"flake8",
 	"mypy",
 	"isort"
 ]
